+++ packages/sdk-core/test/0_framework.test.ts
@@ -44,6 +44,7 @@ describe("Framework Tests", async () => {
             try {
                 await Framework.create({ provider: deployer.provider!, chainId: null as any });
             } catch (err: any) {
+                expect(err.name).to.equal("SFError");
                 expect(err.message).to.equal(
                     "Framework Initialization Error: You must input chainId."
                 );
@@ -59,6 +60,7 @@ describe("Framework Tests", async () => {
                     provider: deployer.provider!,
                 });
             } catch (err: any) {
+                expect(err.name).to.equal("SFError");
                 expect(err.message).to.equal(
                     "Network Mismatch Error: Your provider network chainId is: " +
                         chainId +
@@ -75,6 +77,7 @@ describe("Framework Tests", async () => {
                     chainId: MATIC_CHAIN_ID,
                 } as any);
             } catch (err: any) {
+                expect(err.name).to.equal("SFError");
                 expect(err.message).to.equal(
                     "Framework Initialization Error: You must pass in a provider, an injected web3.js or ethers.js instance when initializing the framework."
                 );
@@ -92,6 +95,7 @@ describe("Framework Tests", async () => {
                     protocolReleaseVersion: "test",
                 });
             } catch (err: any) {
+                expect(err.name).to.equal("SFError");
                 expect(err.message).to.equal(
                     "Framework Initialization Error: You must input your own resolver address if you use an unsupported network."
                 );
@@ -113,6 +117,7 @@ describe("Framework Tests", async () => {
                     protocolReleaseVersion: "test",
                 });
             } catch (err: any) {
+                expect(err.name).to.equal("SFError");
                 expect(err.message).to.contain(
                     "Framework Initialization Error: There was an error initializing the framework"
                 );