+++ packages/sdk-core/CHANGELOG.md
@@ -11,7 +11,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Breaking
 - Support for: `rinkeby`, `ropsten`, `kovan`, `optimism-kovan` and `arbitrum-rinkeby` removed
 - Don't throw `SFError` when executing `Operation` or `BatchCall`; let the original error bubble up
+
+### Fixes
 - Serialize a much smaller version of the cause in `SFError` (only `name`, `message`, `code`)
+- Change `SFError.name` from "Error" to "SFError"
 # [0.5.4] - 2022-08-19
 
 ### Fixes